requirements.txt
# Created by 'make update-requirements-txt'. DO NOT EDIT!
click==8.0.4
PyYAML==6.0
typer==0.4.0
